ID Split - building lib-tempo
