+++ src/lib-tempo/include/tempo/entity/PlayerRemote.hpp
@@ -5,7 +5,7 @@
 #include <anax/Component.hpp>
 
 #include <tempo/component/ComponentGridMotion.hpp>
-#include <tempo/entity/ID.hpp>
+#include <tempo/entity/SystemID.hpp>
 #include <tempo/entity/SystemQID.hpp>
 #include <tempo/network/queue.hpp>
 #include <tempo/time.hpp>
