+++ src/lib-tempo/include/tempo/entity/PlayerRemoteServer.hpp
@@ -5,7 +5,7 @@
 #include <anax/Component.hpp>
 
 #include <tempo/component/ComponentGridMotion.hpp>
-#include <tempo/entity/ID.hpp>
+#include <tempo/entity/SystemID.hpp>
 #include <tempo/entity/SystemQID.hpp>
 #include <tempo/network/queue.hpp>
 #include <tempo/network/server.hpp>
