+++ src/lib-tempo/include/tempo/entity/SystemID.hpp
@@ -1,5 +1,5 @@
-#ifndef TEMPO_ENTITY_ID_HPP
-#define TEMPO_ENTITY_ID_HPP
+#ifndef TEMPO_SYSTEM_ID_HPP
+#define TEMPO_SYSTEM_ID_HPP
 
 #include <anax/System.hpp>
 #include <anax/Component.hpp>
@@ -7,24 +7,9 @@
 #include <iostream>
 #include <stdexcept>
 
+#include<tempo/entity/ComponentID.hpp>
+
 namespace tempo{
-
-extern int globalIDCounter;
-extern std::map<int, anax::Entity> id_map;
-	
-struct ComponentID : anax::Component {
-	int instance_id;
-	int type_id;
-
-	//only for server
-	ComponentID(int eid);
-
-	//client
-	ComponentID(int tid, int eid);
-
-	//possibly unnecessary
-	~ComponentID();
-};
 
 struct SystemID : anax::System<anax::Requires<ComponentID>> {
 	std::map<int, anax::Entity> id_map;
